requirements.txt
ansible-vault-win==1.0.0
Django==4.1.7
envparse==0.2.0
psycopg2-binary==2.9.5
gunicorn==20.1.0
djangorestframework==3.14.0
django-extensions==3.2.1
django-filter==23.1
social-auth-app-django==5.1.0
pydantic==1.10.7
requests==2.28.2
colorama==0.4.6
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
tomli==2.0.1
Faker==18.4.0
factory-boy==3.2.1
python-dateutil==2.8.2
six==1.16.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
